backend/pyproject.toml
[project]
name = "Hibou-Interface-Backend"
version = "0.1.0"
description = "Backend of the Hibou interface"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "bcrypt==4.0.1",
    "fastapi[standard]>=0.135.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.7.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy[asyncio]>=2.0.48",
    "pyzmq>=26.0.0",
    "numpy>=2.0.0",
    "opencv-python-headless>=4.10.0",
]
